Add srgb/srgb texture and fbo write support on OSX.  Log the text of the shader and whether vertex/fragment shaders failed.
This is a breaking change if code on OSX is using the SRGB_TO_RGB() macros in the fragment shaders.
Want to eliminate use of the srgb macros at the end of every shader, since those don't work for blending.
Also srgb formats are needed for correct srgb reads of albedo textures.