fix snprintf usage in ToString functions
snprintf returns the number of characters that would have been printed
if the size were unlimited, so you can't use the returned size when
copying memory because if the string was truncated you are now reading
out of bounds from the source buffer.

Also, snprintf already subtracts 1 from the buffer size to ensure space
for a 0 terminator.